requirements.txt
pygame>=2.1.2
PyYAML~=6.0
